requirements.txt
absl-py==2.3.1
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
attrs==25.4.0
cachetools==6.2.1
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
Django==5.2.7
djangorestframework==3.16.1
docstring_parser==0.17.0
filelock==3.20.0
flatbuffers==25.9.23
fonttools==4.60.1
fsspec==2025.9.0
google-api-core==2.28.0
google-auth==2.41.1
google-cloud-aiplatform==1.122.0
google-cloud-bigquery==3.38.0
google-cloud-core==2.4.3
google-cloud-resource-manager==1.15.0
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-genai==1.46.0
google-resumable-media==2.7.2
googleapis-common-protos==1.71.0
grpc-google-iam-v1==0.14.3
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jax==0.7.1
jaxlib==0.7.1
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.7
mediapipe==0.10.21
ml_dtypes==0.5.3
mpmath==1.3.0
mysqlclient==2.2.7
networkx==3.5
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pandas==2.3.3
pillow==12.0.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scipy==1.16.2
sentencepiece==0.2.1
shapely==2.1.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.3
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
torch==2.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websockets==15.0.1

